requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
virtualenv==16.7.9
whitenoise==5.0.1
djangorestframework==3.11.0
django-filter==2.2.0
kaggle==1.5.6
requests==2.22.0
django-background-tasks
